:hammer: Extract system tests utilities in dedicated directory
